Moved ApplicationUser.cs to Shared project. Configured database for IdentityServer. Reset database for Examplium.Server.
