+++ Examplium.IdentityServer/Examplium.IdentityServer.csproj
@@ -4,6 +4,11 @@
     <TargetFramework>net7.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
+    <UserSecretsId>f2c3395a-42c7-4ddb-a5d1-6ce961fdb93b</UserSecretsId>
   </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Examplium\Shared\Examplium.Shared.csproj" />
+  </ItemGroup>
 
 </Project>
