+++ Examplium.IdentityServer/SecretsTemplate.txt
@@ -1,0 +1,4 @@
+﻿{
+    "DefaultDatabaseConnection": "Server=(localdb)\\mssqllocaldb;Database=Examplium.IdentityServer;Trusted_Connection=True;MultipleActiveResultSets=true",
+    "AuthorizationServerUrl": "[Auth/Identity Server address]"
+}
