+++ Examplium/Server/Areas/Identity/Pages/Shared/_LoginPartial.cshtml
@@ -1,5 +1,5 @@
-﻿@using Microsoft.AspNetCore.Identity
-@using Examplium.Server.Models
+﻿@using Examplium.Shared.Models.Identity
+@using Microsoft.AspNetCore.Identity
 @inject SignInManager<ApplicationUser> SignInManager
 @inject UserManager<ApplicationUser> UserManager
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
