+++ Examplium/Server/Examplium.Server.csproj
@@ -30,4 +30,9 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Data\Migrations\" />
+    <Folder Include="Models\" />
+  </ItemGroup>
+
 </Project>
