+++ Examplium/Server/SecretsTemplate.txt
@@ -1,4 +1,4 @@
 ﻿{
-    "DefaultDatabaseConnection": "Server=(localdb)\\mssqllocaldb;Database=aspnet-Examplium.Server-46763ee3-2951-4fc0-b6d2-486b57b9d2fb;Trusted_Connection=True;MultipleActiveResultSets=true",
+    "DefaultDatabaseConnection": "Server=(localdb)\\mssqllocaldb;Database=Examplium.Server;Trusted_Connection=True;MultipleActiveResultSets=true",
     "AuthorizationServerUrl": "[Auth/Identity Server address]"
 }
